fix init typeorm

+++ src/api/command/commands/delete-account.command.ts
@@ -11,7 +11,17 @@ export class DeleteAccountCommand extends BaseCommand {
     super();
   }
 
-  async execute({ params }: TypedCommandDescriptor): Promise<CommandResult> {
+  async execute(typedCommandDescriptor: TypedCommandDescriptor): Promise<CommandResult> {
+    const { params } = typedCommandDescriptor;
+
+    const helpResult = await super.execute(typedCommandDescriptor);
+
+    if (helpResult) {
+      return helpResult;
+    }
+
+    delete params['help'];
+
     await this.accountService.delete(params.id);
 
     return { result: `Account with id=${params.id} deleted` };
