+++ src/api/command/commands/delete-bank.command.ts
@@ -11,7 +11,17 @@ export class DeleteBankCommand extends BaseCommand {
     super();
   }
 
-  async execute({ params }: TypedCommandDescriptor): Promise<CommandResult> {
+  async execute(typedCommandDescriptor: TypedCommandDescriptor): Promise<CommandResult> {
+    const { params } = typedCommandDescriptor;
+
+    const helpResult = await super.execute(typedCommandDescriptor);
+
+    if (helpResult) {
+      return helpResult;
+    }
+
+    delete params['help'];
+
     await this.bankService.delete(params.id);
 
     return { result: `Bank with id=${params.id} deleted` };
